show project assets in paid bar
